# Introduce more typechecks and refactor