--- a/src/brstm/.gitignore
+++ b/src/brstm/.gitignore
@@ -1,3 +1,4 @@
 dist/
 node_modules/
 .cache/
+!tests/*.brstm