--- a/src/brstm/.gitignore
+++ b/src/brstm/.gitignore
@@ -1,4 +1,3 @@
 dist/
 node_modules/
 .cache/
-!tests/*.brstm