--- a/src/brstm/jsconfig.json
+++ b/src/brstm/jsconfig.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "strict": true
+  },
+  "exclude": ["tests/"]
+}