--- a/src/brstm/tests/assets/info.md
+++ b/src/brstm/tests/assets/info.md
@@ -1,0 +1,3 @@
+# Source
+
+https://commons.wikimedia.org/wiki/File:Malfidus_news_loop.ogg