--- a/src/brstm/tests/helper.js
+++ b/src/brstm/tests/helper.js
@@ -1,0 +1,25 @@
+//@ts-check
+
+const assert = require('assert').strict;
+
+function test(testName, block) {
+  try {
+    block();
+    console.log(`[OK] ${testName}`)
+  } catch (e) {
+    console.error(`[FAIL] ${testName}: ${e.message}`, e);
+  }
+}
+
+function expect(actualValue) {
+  return {
+    toBe: (expectedValue) => {
+      assert.equal(actualValue, expectedValue);
+    },
+  };
+}
+
+module.exports = {
+  test,
+  expect,
+};