--- a/src/brstm/tests/index.test.js
+++ b/src/brstm/tests/index.test.js
@@ -24,10 +24,12 @@ test('Get all samples', () => {
   expect(allSamples[0][15000]).toBe(-435);
   expect(allSamples[0][25000]).toBe(-627);
   expect(allSamples[0][35000]).toBe(-1951);
+  expect(allSamples[0][856000]).toBe(-581);
   expect(allSamples[1][5000]).toBe(-64);
   expect(allSamples[1][15000]).toBe(-679);
   expect(allSamples[1][25000]).toBe(-707);
   expect(allSamples[1][35000]).toBe(-1183);
+  expect(allSamples[1][856000]).toBe(-219);
 });
 
 test('Get partial samples', () => {
@@ -40,8 +42,46 @@ test('Get partial samples', () => {
   expect(samples[0][15000]).toBe(-435);
   expect(samples[0][25000]).toBe(-627);
   expect(samples[0][35000]).toBe(-1951);
+  expect(samples[0][856000]).toBe(-581);
   expect(samples[1][5000]).toBe(-64);
   expect(samples[1][15000]).toBe(-679);
   expect(samples[1][25000]).toBe(-707);
   expect(samples[1][35000]).toBe(-1183);
+  expect(samples[1][856000]).toBe(-219);
+});
+
+test('Get partial samples (2)', () => {
+  const brstm = new Brstm(testFile.buffer);
+  const samples = brstm.getSamples(0, 15001);
+  expect(samples.length).toBe(2);
+  expect(samples[0].length).toBe(15001);
+
+  expect(samples[0][5000]).toBe(478);
+  expect(samples[0][15000]).toBe(-435);
+  expect(samples[1][5000]).toBe(-64);
+  expect(samples[1][15000]).toBe(-679);
+});
+
+test('Get partial samples (3)', () => {
+  const brstm = new Brstm(testFile.buffer);
+  const offset = 856000;
+  const samples = brstm.getSamples(offset, 1);
+  expect(samples.length).toBe(2);
+  expect(samples[0].length).toBe(1);
+
+  expect(samples[0][856000 - offset]).toBe(-581);
+  expect(samples[1][856000 - offset]).toBe(-219);
+});
+
+test('Get partial samples (4)', () => {
+  const brstm = new Brstm(testFile.buffer);
+  const offset = 25000;
+  const samples = brstm.getSamples(offset, 10001);
+  expect(samples.length).toBe(2);
+  expect(samples[0].length).toBe(10001);
+
+  expect(samples[0][25000 - offset]).toBe(-627);
+  expect(samples[0][35000 - offset]).toBe(-1951);
+  expect(samples[1][25000 - offset]).toBe(-707);
+  expect(samples[1][35000 - offset]).toBe(-1183);
 });