- More type checks using JSDoc. [Powered by TypeScript](https://www.typescriptlang.org/docs/handbook/type-checking-javascript-files.html#supported-jsdoc)
- Refactor: Rename `getBuffer` to `getSamples`, old method still exist for compatibility
- Refactor: Extract common logic of `getSamples` and `getAllSamples` so they share the same codes.
- Introduced tests using a very basic test runner

:warning: This might introduce some bugs on both correctness and performance. I'll be testing this part for some time first before merging

TODOs

- [x] Measure performance before/after this PR
- [x] Check other BRSTM files whether they are still playing correctly or not

Update:

Taken 10 samples on Firefox 76.0.1 (64-bit) under Windows 10 1909. Computer processor is Intel Core i7 8750H @ 2.2 GHz. BRSTM file used for this testing is [The Afternoon Streets](https://smashcustommusic.net/song/16232)

  | This branch (ms) | master (ms)
-- | -- | --
Mean | 173.9 | 171.4
Median | 175.5 | 171
Standard deviation (sample) | 8.279157 | 11.55855

A slowdown of 1% was observed (but I think that it is not significant). This might be explained by the more function calls need to be made